edutrack_pro/requirements.txt
Flask
bcrypt
PyJWT
requests
beautifulsoup4
streamlit
matplotlib